+++ confik/examples/bigdecimal.rs
@@ -1,29 +1,27 @@
+use std::str::FromStr;
+
+use bigdecimal::BigDecimal;
 use confik::{Configuration, TomlSource};
-use indoc::indoc;
-use secrecy::{ExposeSecret as _, SecretString};
+use indoc::formatdoc;
 
 #[derive(Configuration, Debug)]
 struct Config {
-    secret_field: SecretString,
+    big_decimal: BigDecimal,
 }
 
 fn main() {
-    let toml = indoc! {r#"
-        secret_field = "ProtectedSecret"
+    let big_decimal = "1.414213562373095048801688724209698078569671875376948073176679737990732478462107038850387534327641573";
+    let toml = formatdoc! {r#"
+        big_decimal = "{big_decimal}"
     "#};
 
     let config = Config::builder()
-        .override_with(TomlSource::new(toml).allow_secrets())
+        .override_with(TomlSource::new(toml))
         .try_build()
         .expect("Failed to parse config");
 
     assert_eq!(
-        format!("{config:?}"),
-        "Config { secret_field: Secret([REDACTED alloc::string::String]) }",
+        config.big_decimal,
+        BigDecimal::from_str(big_decimal).unwrap()
     );
-    assert_eq!(
-        format!("{:?}", config.secret_field),
-        "Secret([REDACTED alloc::string::String])",
-    );
-    assert_eq!(config.secret_field.expose_secret(), "ProtectedSecret");
 }