+++ confik/tests/third_party.rs
@@ -66,3 +66,63 @@ mod secrecy {
         assert_eq!(config.secret_string.expose_secret(), "SeriouslySecret");
     }
 }
+
+#[cfg(feature = "bigdecimal")]
+mod bigdecimal {
+    use std::str::FromStr;
+
+    use bigdecimal::BigDecimal;
+    use confik::{Configuration, Error, TomlSource};
+    use indoc::formatdoc;
+
+    #[derive(Configuration, Debug)]
+    struct Config {
+        big_decimal: BigDecimal,
+    }
+
+    #[test]
+    fn bigdecimal() {
+        let big_decimal = "1.414213562373095048801688724209698078569671875376948073176679737990732478462107038850387534327641573";
+        let toml = formatdoc! {r#"
+            big_decimal = "{big_decimal}"
+        "#};
+
+        let config = Config::builder()
+            .override_with(TomlSource::new(toml))
+            .try_build()
+            .expect("Failed to parse config");
+
+        assert_eq!(
+            config.big_decimal,
+            BigDecimal::from_str(big_decimal).unwrap()
+        );
+    }
+
+    #[test]
+    fn bigdecimal_missing_err_propagation() {
+        let toml = formatdoc! {r#"
+            big_decimal = ""
+        "#};
+
+        let config_parsing_err = Config::builder()
+            .override_with(TomlSource::new(toml))
+            .try_build();
+        match config_parsing_err {
+            Ok(_) => {
+                panic!("Expected parsing error");
+            }
+            Err(err) => match err {
+                Error::Source(source_err, _config) => {
+                    assert!(source_err
+                        .to_string()
+                        .contains("Failed to parse empty string"));
+                    assert!(source_err.to_string().contains("big_decimal"));
+                }
+
+                _ => {
+                    panic!("Expected MissingValue error");
+                }
+            },
+        }
+    }
+}